NO-ISSUE: Fix mismatching marshal and unmarshal of install config overrides
Based on bug found during scale testing where the host finishes install
but the cluster doesn't. Culprit is that the install config overrides
were marshaled as JSON, but unmarshaled as YAML. This led to missing
overrides for capabilities, and that led to the console operator being
seen as enabled in the service, but disabled in the installer controller.

https://redhat-internal.slack.com/archives/CUPJTHQ5P/p1708377415618139